# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Family Health Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

/**
 * @nullable
 */
export type UserGender = (typeof UserGender)[keyof typeof UserGender] | null;

export const UserGender = {
  male: "male",
  female: "female",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface User {
  id: number;
  username: string;
  displayName: string;
  role: UserRole;
  /** @nullable */
  gender: UserGender;
  status: UserStatus;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export type RegisterBodyGender =
  (typeof RegisterBodyGender)[keyof typeof RegisterBodyGender];

export const RegisterBodyGender = {
  male: "male",
  female: "female",
} as const;

export interface RegisterBody {
  username: string;
  password: string;
  displayName: string;
  gender: RegisterBodyGender;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export type CreateUserBodyGender =
  (typeof CreateUserBodyGender)[keyof typeof CreateUserBodyGender];

export const CreateUserBodyGender = {
  male: "male",
  female: "female",
} as const;

export type CreateUserBodyStatus =
  (typeof CreateUserBodyStatus)[keyof typeof CreateUserBodyStatus];

export const CreateUserBodyStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface CreateUserBody {
  username: string;
  password: string;
  displayName: string;
  role: CreateUserBodyRole;
  gender: CreateUserBodyGender;
  status?: CreateUserBodyStatus;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export type UpdateUserBodyGender =
  (typeof UpdateUserBodyGender)[keyof typeof UpdateUserBodyGender];

export const UpdateUserBodyGender = {
  male: "male",
  female: "female",
} as const;

export type UpdateUserBodyStatus =
  (typeof UpdateUserBodyStatus)[keyof typeof UpdateUserBodyStatus];

export const UpdateUserBodyStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface UpdateUserBody {
  displayName?: string;
  role?: UpdateUserBodyRole;
  gender?: UpdateUserBodyGender;
  status?: UpdateUserBodyStatus;
}

export interface ResetPasswordBody {
  newPassword: string;
}

export interface WeightEntry {
  id: number;
  userId: number;
  weightKg: number;
  recordedAt: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface CreateWeightEntryBody {
  weightKg: number;
  recordedAt: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  memberId?: number | null;
}

export interface UpdateWeightEntryBody {
  weightKg?: number;
  recordedAt?: string;
  /** @nullable */
  note?: string | null;
}

export interface LengthEntry {
  id: number;
  userId: number;
  lengthCm: number;
  recordedAt: string;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface CreateLengthEntryBody {
  lengthCm: number;
  recordedAt: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  memberId?: number | null;
}

export interface UpdateLengthEntryBody {
  lengthCm?: number;
  recordedAt?: string;
  /** @nullable */
  note?: string | null;
}

export interface PeriodEntry {
  id: number;
  userId: number;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  numberOfDays: number | null;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface CreatePeriodEntryBody {
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  numberOfDays?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface UpdatePeriodEntryBody {
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  numberOfDays?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface HealthSummary {
  /** @nullable */
  latestWeight: number | null;
  /** @nullable */
  weightTrend: number | null;
  /** @nullable */
  latestLength: number | null;
  weightEntriesCount: number;
  lengthEntriesCount: number;
  /** @nullable */
  lastPeriodStart: string | null;
  /** @nullable */
  nextPeriodEstimate: string | null;
  /** @nullable */
  avgCycleLength: number | null;
}

export interface AdminOverview {
  totalUsers: number;
  activeUsers: number;
  pendingApprovals: number;
  suspendedUsers: number;
  totalWeightEntries: number;
  totalLengthEntries: number;
  recentRegistrations: User[];
}

export type ListUsersParams = {
  status?: ListUsersStatus;
};

export type ListUsersStatus =
  (typeof ListUsersStatus)[keyof typeof ListUsersStatus];

export const ListUsersStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export type ListWeightEntriesParams = {
  /**
   * Family member user ID (if viewing own family member)
   * @nullable
   */
  memberId?: number | null;
};

export type ListLengthEntriesParams = {
  /**
   * Family member user ID
   * @nullable
   */
  memberId?: number | null;
};