# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family Health Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration submitted, pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account suspended or pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /length:
    get:
      operationId: listLengthEntries
      tags: [length]
      summary: List length/height entries for the current user
      parameters:
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
          description: Family member user ID
      responses:
        "200":
          description: List of length entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LengthEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createLengthEntry
      tags: [length]
      summary: Add a length/height entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLengthEntryBody"
      responses:
        "201":
          description: Length entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LengthEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /length/{id}:
    patch:
      operationId: updateLengthEntry
      tags: [length]
      summary: Update a length/height entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLengthEntryBody"
      responses:
        "200":
          description: Updated length entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LengthEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLengthEntry
      tags: [length]
      summary: Delete a length/height entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /period:
    get:
      operationId: listPeriodEntries
      tags: [period]
      summary: List period entries for the current user (private)
      responses:
        "200":
          description: List of period entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodEntry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (not available for this user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPeriodEntry
      tags: [period]
      summary: Add a period entry (private)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePeriodEntryBody"
      responses:
        "201":
          description: Period entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /period/{id}:
    patch:
      operationId: updatePeriodEntry
      tags: [period]
      summary: Update a period entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePeriodEntryBody"
      responses:
        "200":
          description: Updated period entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PeriodEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePeriodEntry
      tags: [period]
      summary: Delete a period entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, user]
        gender:
          type: ["string", "null"]
          enum: [male, female, null]
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - role
        - gender
        - status
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        displayName:
          type: string
        gender:
          type: string
          enum: [male, female]
      required:
        - username
        - password
        - displayName
        - gender
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
    LengthEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        lengthCm:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - lengthCm
        - recordedAt
        - note
        - createdAt
    CreateLengthEntryBody:
      type: object
      properties:
        lengthCm:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        memberId:
          type: ["integer", "null"]
      required:
        - lengthCm
        - recordedAt
    UpdateLengthEntryBody:
      type: object
      properties:
        lengthCm:
          type: number
          format: float
        recordedAt:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
    PeriodEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        numberOfDays:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - startDate
        - endDate
        - numberOfDays
        - note
        - createdAt
    CreatePeriodEntryBody:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        numberOfDays:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
      required:
        - startDate
    UpdatePeriodEntryBody:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        numberOfDays:
          type: ["integer", "null"]
        note:
          type: ["string", "null"]
...