# lib/api-zod/src/generated/types/periodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Family Health Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface PeriodEntry {
  id: number;
  userId: number;
  startDate: Date;
  /** @nullable */
  endDate: Date | null;
  /** @nullable */
  numberOfDays: number | null;
  /** @nullable */
  note: string | null;
  createdAt: Date;
}