# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Family Health Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterBodyGender } from "./registerBodyGender";

export interface RegisterBody {
  username: string;
  password: string;
  displayName: string;
  gender: RegisterBodyGender;
}