# lib/api-zod/src/generated/types/updateLengthEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Family Health Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLengthEntryBody {
  lengthCm?: number;
  recordedAt?: Date;
  /** @nullable */
  note?: string | null;
}