# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Family Health Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UserGender } from "./userGender";
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  username: string;
  displayName: string;
  role: UserRole;
  /** @nullable */
  gender: UserGender;
  status: UserStatus;
  createdAt: Date;
}